+++ src/App.jsx
@@ -1,8 +1,10 @@
 import Landing from "./Landing.jsx";
+import Location from "./Routes/Location.jsx";
 
 const App = () => {
   return (
     <Landing />
+    // <Location />
   );
 };
