+++ src/Routes/BlogPage.jsx
@@ -1,6 +1,6 @@
 
 import Navbar from '../components/Navbar'
-import Footer from '../components/Footer'
+import Footer from "../components/footers/Footer";
 
 function BlogPage()  {
   return (
