+++ src/Routes/LocalEventsPage.jsx
@@ -1,6 +1,5 @@
-
 import Navbar from '../components/Navbar'
-import Footer from '../components/Footer'
+import Footer from '../components/footers/Footer'
 
 function LocalEventsPage()  {
   return (
